E-Commerce-Website-LifestyleMart-main/E-Commerce-Website-LifestyleMart-main/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
SQLAlchemy>=2.0.36
PyMySQL==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0
Pillow>=10.0.1
marshmallow==3.20.1
gunicorn==21.2.0
requests
